backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
yfinance==0.2.18
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
python-multipart==0.0.6
python-dotenv==1.0.0